requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
boto3==1.36.12
botocore==1.36.12
celery==5.4.0
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
distlib==0.3.9
Django==5.1.5
django-storages==1.14.4
djangorestframework==3.15.2
filelock==3.16.1
jmespath==1.0.1
kombu==5.4.2
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==10.4.0            # <— was 11.1.0; keep <11 for gradio 4.44.x
pipenv==2024.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psycopg2==2.9.10
pycparser==2.22
PyMuPDF==1.25.2
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
virtualenv==20.28.0
wcwidth==0.2.13

faiss-cpu~=1.10.0
# fitz~=0.0.1.dev2      # <— remove this; it's not PyMuPDF
sentence-transformers~=3.4.1
pathlib~=1.0.1
requests~=2.32.3
gradio~=4.44.0          # works with Pillow <11
# Optional: enable reranking with HF transformers (uncomment to use)
transformers>=4.45.0,<5.0.0
torch>=2.2
docx2txt~=0.8
